requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
google-generativeai==0.3.2
edge-tts==6.1.9
vosk==0.3.45
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.4
websockets==11.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
Jinja2==3.1.2
Werkzeug==2.3.7